Refactored code..

+++ src/main/java/com/affirm/loan/core/Facility.java
@@ -11,7 +11,7 @@ public class Facility {
 
     int id;
     double interestRate;
-    double amountInCents;
+    long amountInCents;
     double calculatedYield;
 
     Bank bank;
